Refactor the entire codebase!
Move all dataset dependent functions within the dataset class
